[owners] Remove rudymathu@google.com from tools/bootserver/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://fuchsia-review.googlesource.com/q/rudymathu@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Ibcd89d867f6998dfbaaab0c06be46259e3e47a01
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1365943
Fuchsia-Auto-Submit: Owners-Cleanup-Prod <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Commit-Queue: Owners-Cleanup-Prod <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>