[assembly] Indicate "test ramdisk"-ness in product bundle
A follow-on from fxrev.dev/1352052, this change forwards the ramdisk to
the product bundle as having type "test-ramdisk" for assembly mode
"TestRamdisk". This is a signal to downstream tooling consuming such
product bundles and images (e.g., `ffx emu`) to also treat these
ramdisks as opaque, final objects, not to be modified.

A follow-on change to this one will be made to `ffx emu` to handle such
product bundles.

Bug: 432535854, 439032069
Change-Id: I13a89b5e8448f226b9d15bea20695aa83cd2364f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1364597
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>