[driver_manager] Use transient collections
This change is a major refactor of how driver manager and component
framework interact with each other. The main driver for this is a desire
to switch away from the single run driver component, and instead use the
standard transient type. This will also allow us to reuse existing
component framework features like reload in the future instead of
keeping our own standalone implementation around.

The change introduces some more states in the node shutdown flow, and
separates the concept of stopping a component from destroying it.

It also moves away from the token in startup handles for validating
start requests, and instead uses the component framework's GetMoniker
method in order to find which node a start request belongs to.

Change-Id: Ic7d578670d3b804db27749c508ff6d2876e0b4e1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1328808
Commit-Queue: Novin Changizi <novinc@google.com>
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Reviewed-by: Ali Zhang <alizhang@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>