[kernel][vm] Move cache_policy_ to VmObject
Both VmObjectPaged and VmObjectPhysical define a mapping cache policy.
Although they have different rules around when it can be modified, the
storage and retrieving is common.

The primary benefit here is that the VmObject has some spare padding
space and moving this is the first step to saving memory in
VmObjectPaged.

Change-Id: Ice9b4185f03f1cb048ca748129ec9c77c905e0aa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1361832
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>