[ffx][emu] Support opaque, unmodifiable, test ramdisks
This change is a follow-on from fxrev.dev/1364597 that surfaces the
notion of a test ramdisk to the product bundle, which basically
indicates a (QEMU) ramdisk intended for testing that should be booted
as-is (i.e., without modification). The bug in question tracks
supporting a boot test that features a non-ZBI ramdisk, but in practice
a proper ZBI may be passed as a test ramdisk too. We add straightforward
support for this in ffx emu, really just suppressing the call to modify
the input ZBI to embed some select things (like an SSH public key).

Bug: 432535854
Change-Id: I7ad87325e479c2edcc3ad9f6ea965e54147257e6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1364599
Reviewed-by: Steven Grady <slgrady@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>