[firmware] add --split arg to android_boot_image.py
It's useful for testing and development to be able to easily split
concatenated boot images. Currently the process is to run this tool and
then manually split images e.g. with `dd`; this CL adds a `--split` arg
to this tool to simplify it.

Bug: b/434935097
Change-Id: I7ab28590c843aee4ee3fff3cd9f0c7126d928532
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1366616
Reviewed-by: Yecheng Zhao <zyecheng@google.com>
Fuchsia-Auto-Submit: David Pursell <dpursell@google.com>
Commit-Queue: David Pursell <dpursell@google.com>