[syslog] Remove the legacy syslog library from allowlist
This removes the legacy syslog library from the shared library
allowlist.  Technically, this library can still be used as it is still
in currently supported SDK versions. If we continue to allow drivers to
link with this library, current driver limitations mean that I have to
keep the library working at HEAD because that's the version that is made
available to drivers (regardless of the version they might link to via
the SDK).

I have done an audit, and I don't believe there are any drivers that
still link to the legacy logging library, so I think that we can safely
remove this library now.

This will then allow me to remove the library at NEXT which will in turn
allow me to remove ConnectStructured at NEXT.

Bug: 377963307
Change-Id: I753d7a9d42e527ad00465e32ec461a6faa7b4463
Run-All-Tests: true
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1360512
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Commit-Queue: Chris Suter <csuter@google.com>