[component_manager][services] Transition to idle on any idle
A component will not be able to view their service directory until
all currently known providers have transitioned to idle.

This is because we would like to prepopulate the services instances
as much as possible before giving it to a component.

There is a bug here, which is that we wait for each directory in
the service path to exist, e.g: "/svc/my.service.name.Service".

If we hit an idle while walking this path, we should move into
the idle state because we've gone as far as we can now.

Before this CL, we fired an idle if it existed but stayed in the
waiting state. This could cause a component to idle foreever
if this idle was hit before `add_entries_from_instance` was called.

Bug: b/439108465
Change-Id: I23bf4614a776d15557a51f06ca992374c80a24fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1366596
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>