[fake-display] Store DriverLayers for applied configuration.
This change replaces the `applied_image_id_` and
`applied_fallback_color_` fields in FakeDisplay with a vector of
DriverLayers which contains the layer information.

This makes it possible for FakeDisplay to support compositing multiple
layers using software or Vulkan-based compositor when capturing the
screen contents.

Bug: 124683
Test: fake-display-unittests
Change-Id: I36b176c9e1af835ac1979a198630325654ea626f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/882511
Commit-Queue: Yilong Li <liyl@google.com>
Reviewed-by: Josh Gargus <jjosh@google.com>