[kernel][phys] Remove TODO around physmap dimensions
While it might be nice to replace the duplicated variables giving the
physmap dimensions with a temporary pointer to the physmap PhysVmar,
that's difficult to do with the sorting of VMARs that occurs in
constructing the hand-off, which is itself nice for normalization's and
early boot logging's sake. We remove the TODO and clarify exactly that
with some new comments.

Bug: 42164859
Change-Id: I0fdfd634db6510b332bd0da528ca3d9d3821912f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1367342
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>