[kernel][vm] Redesign PageQueues synchronization
This change redesigns how the different threads that are part of, and
interact with, the PageQueues synchronize with each other.

The overall strategy of this change is to have all the relevant state
be explicit, transition under the lock, and only use events to notify
threads that they might have work to do.

A consequence of this shift in strategy is that specific actions,
such as performing aging, do not have have to be done specifically on
the MruThread, rather the MruThread is just a persistent thread ready
to perform aging, but any thread in the system could potentially
perform it if required.

Having other threads perform aging is not the specific goal of this
change, but the fact that it is possible has the important side
effect of making it so that any thread can synchronize with aging.
Importantly this allows reclamation, via PeekIsolate and the
SynchronizeWithAging methods, to reliably perform the minimum
synchronization with the MruThread, without necessarily needing to
wait for a full aging step to complete. In particular
SynchronizeWithAging *can* directly perform the increment of the MRU
generation if necessary, but never has to wait if the MRU cannot be
incremented, since if the MRU cannot be incremented then this implies
there are LRU items to process and reclaim.

The goal of avoiding SynchronizeWithAging from unnecessarily blocking
behind MRU+LRU processing could have been achieved in other ways, but
this would have layered additional complexity onto what is already a
complex synchronization system. Even without SynchronizeWithAging the
synchronization in this commit is simpler and easier to reason about.

Bug: 423900604
Change-Id: Ic1ad4d395aaff2e98bd444ef12b9ac7c23709a42
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1354224
Reviewed-by: Ethan Williams <ethanws@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>