[flatland][display] Always allow color layers regardless of their sizes.
Previously, DisplayCompositor rejected color layers for
direct-to-display composition unless it's a background layer with
screen size.

For devices that didn't support non-background color layers (for
example, amlogic-display), this saved a CheckConfig() call; but it also
made it impossible for other devices to use direct-to-display for such
configs.

This change removes the constraints from DisplayCompositor. The display
Coordinator and Engine drivers will be responsible for checking if the
color layers are supported in `CheckConfig()`.

Theoretically, this may affect Scenic performance as it requires an
extra CheckConfig() when the layer config changes. In practice,
production devices are unlikely to have a client creating non-background
color layers, so the performance should be minimally affected.

Test: Manually verified that flatland-pixeltest-general can use
      multi-layer composition.
Bug: 444575885
Bug: 42056054
Change-Id: I91aa3b837cde40943d3e713f2b26d423626021e1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1366467
Reviewed-by: Josh Gargus <jjosh@google.com>
Commit-Queue: Yilong Li <liyl@google.com>
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>