[display][software-compositor] Add support for solid color layers.
This change enhances the software compositor by enabling it to render
layers with solid colors, in addition to layers with images. This is
achieved by introducing a `fallback_color` to be used when a layer
does not have an associated image.

Test: software-compositor-tests
Bug: 42075534
Change-Id: Idd4b73ffa4e33248f02ddde0c1356f983bd34cb0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1365588
Reviewed-by: Josh Gargus <jjosh@google.com>
Commit-Queue: Yilong Li <liyl@google.com>
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>