[memory_monitor] Fix erroneous infinite recursion
When processing a capture of the current memory usage, both
`memory_monitor`s iterate through VMOs' descendance graph to share
memory attribution between a parent and their children. The graph used
to be acyclic, but this is no longer true (at least `IOBuffer`'s store
their own id as their parent id as an implementation detail).

In `memory_monitor`, `CaptureMaker::ReallocateDescendants` is recursive
and does not detect cycles, which leads to unbounded recursion and
eventually overflows the stack, causing memory errors and leading to
the crash of memory_monitor.

In `memory_monitor2`, `processing::attribute_vmos` iterates through a
VMO's parents; it does not do so via recursion so the stack should be
safe from overflows and memory corruptions, but it can lead to
unbounded iteration time as well (which is likely to stall it).

This change ensures this kind of cycles (self-referential VMOs) get
detected and preemptively cut in both `memory_monitor`'s.

Bug: 436248158
Change-Id: Ib4fa80aa178d1de5b1d2cacd09e61380bd0e0641
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1367157
Reviewed-by: Étienne J. Membrives <etiennej@google.com>
Commit-Queue: Guillaume Chelfi <chelfi@google.com>