[starnix] Log the last reboot reason to console-ramoops
After a crash reboot, the body of the ramoops file is propagated to APC.
Log the last reboot reason in the ramoops to make it accessible from the
dashboard.

This will provide access to the raw reboot reason, unlike the Android
system property having a converted reason from the set of Linux reboot
reasons.

Bug: 438833540
Test: ffx shell$ k crash illegal_instruction
  sorrel:/ $ cat /sys/fs/pstore/console-ramoops
  Passed starnix_bootreason_integration_test
Signed-off-by: Dodam Kim <dodam@google.com>
Change-Id: I251cc16a190cf28cab3bdff744587cf18deb11eb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1366261
Reviewed-by: Haines Sy <hainesy@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>