[kernel][vm] k command for orphaned VMOs
`k zx vmos orphaned` dumps memory held on by VMOs that will never be
referenced and should ideally have been freed. Unreachable memory is
expensive to find hence we don't free it today. This command is meant as
a handy tool to determine if any heuristics to find and free up such
memory would be worthwhile.

Currently we estimate two types of orphaned memory:
 - Unreachable memory held in VmCowPages that only have slices
   referencing it but the parent VMO has gone away.
 - Unreachable memory held in VmCowPages that only have user mappings
   but no VMO handles. A few different cases are considered here to also
   ensure that no VMO children can reference the memory.

This is just an initial imperfect version of this command. More changes
can be made in the future for greater accuracy.

Bug: 42069078
Change-Id: Ib3bd452e0f1ed842a596ed8cf80fcd563f11fbe5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1362771
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>