[kernel][uart] Extend kernel UARTs to provide IRQ control.
Prior to this, the only ability a kernel UART implementation had to
control its top level IRQ(s) was during InitInterrupt via a template
injected callback which would allow the specific driver enable the
IRQ(s) at a particular point in the initialization sequence.

Change this to allow UART templates to be expanded with an IrqProvider
(similar to the existing IoProvider).  The common uart::KernelDriver
holds an instance of one of these providers, and can supply it to
various methods which may need more control than a one-time enabling.

A NullIrqProvider (no-op) is supplied in the uart library which does
nothing, and can be used in environments which are not able to use
interrupts (such as physboot).  A KernelIrqProvider is defined in the
kernel's platform/debug layer which can remember the IRQ number, and
also mask/unmask the interrupt in response to user requests.

Note that the ability to _disable_ interrupts in the platform specific
drivers is not being used by any drivers just yet.  This will soon
change in follow-up CLs which add hooks for suspend/resume, and which
will need to both disable and enable interrupts at the proper points
in the suspend/resume sequence.

Bug: 439065000
Change-Id: I095c1a9181a430ca6cbf86b4e15289526bed5a03
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1362595
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>