[profiler] Add sample_id and id to sample_type in perf_event_open
sample_id:  If PERF_SAMPLE_IDENTIFIER is enabled, a 64-bit unique ID is included. This is a duplication of the PERF_SAMPLE_ID id value, but included at the beginning of the sample so parsers can easily obtain the value.

id: If PERF_SAMPLE_ID is enabled, a 64-bit unique ID is included.  If the event is a member of an event group, the group leader ID is returned.  This ID is the same as the one returned by PERF_FORMAT_ID.

Bug: 441344223

Change-Id: I57a08d3e1236128bafa75a9e4db161b6104f782f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1357790
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Jia Li <lizhijia@google.com>
Reviewed-by: Mary Xia <maryxia@google.com>