[trace_processing] Add event filtering to trace conversion
This change introduces the ability to filter trace events by name
during the fxt-to-json conversion process. This is achieved by adding a
new `event_patterns` property to the `MetricsProcessor` class, which
allows each processor to declare the specific events it requires.

The `convert_trace_file_to_json` function is updated to accept a set of
patterns, which are then passed to `trace2json` via the `--pattern`
argument. This allows for more efficient trace processing by avoiding
the conversion of unnecessary events.

This is especially useful for metrics like CPU usage, which only
require a small subset of the events in a trace.

Bug: 437397276
Test: fx test trace_processing_test
Change-Id: I959fe4a7c94fdf7d2f77bf20404a4eb74ac067ab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1364341
Commit-Queue: Chris Masone <cmasone@google.com>
Reviewed-by: Mark Seaborn <mseaborn@google.com>