[starnix] Implement `TimerOps::get_timeline_change_counter`
This is necessary to implement support for `TFD_TIMER_CANCEL_ON_SET`.

The idea is to have a way to observe both the occurrence, and the tally
of UTC timeline changes when they happen. This can then be used to
return the expected `ECANCELED` on a timer read.

This change only puts the support in, but does not yet connect the
wires. That change will be made separately. Each timer optionally gets
its own timeline monitor.

Bug: 297433837
Change-Id: I8b1abeaeb4f345604bbd6748b29e907d3e1d531c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1365796
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>