[alarms] Correct picking the alarm duration for negative durations
Negative durations sometimes happen, when it takes a while to schedule
an alarm with a long duration.

In such situations, wait for the smallest possible time unit. Note,
however that in some devices that smallest possible time unit might
be 1 second.  There is nothing we can do to improve it if the hardware
does not support finer resolutions.

Bug: 445359315
Change-Id: If10f8a1b91e72bb87676223f2c71faf95a3b1f90
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1370802
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Reviewed-by: Guocheng Wei <guochengwei@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>