[feedback] Set network status after successful upload
If an eager upload succeeds, Feedback should use that as an indication
that the network is reachable and attempt to upload any cached reports.

This should only be done for eager uploads because doing so on non-eager
uploads could result in the same report failing to upload many times if
the network reachability is transient.

Fixed: 445454237
Test: `fx test //src/developer/forensics`
Test: `fx shell crasher` while network unavailable; `fx shell crasher`
while available @ 6a270e0bf5059e00

Change-Id: I2f7009ea4dd17e3bdf924ee1cbe88feb61dc8c53
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1369948
Reviewed-by: Alex Pankhurst <pankhurst@google.com>
Commit-Queue: Thomas Gales <tgales@google.com>