[ffx] Introduce "ffx target discover"
Introduce "ffx target discover", a command for populating the
discovery cache. Ffx will use the discovery cache if it exists;
this command actually creates the cache.

By default, "ffx target default" will run in the background,
populating the cache every 60 seconds or so. Various options
support running in the foreground, running in quiet mode, running
with a different period, and stopping the background process.

When invoked when background process is already running, this command
will signal the process, causing it to rerun discovery immediately:
the intended use-case is that when a user knows the cache is out
of date (e.g. because they just unplugged a device), they can cause
the cache to be updated immediately.

Test: comprehensive set of unit tests

Change-Id: I6a6a6964743885a09ef68173b962a969924cec6e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1370045
Commit-Queue: Steven Grady <slgrady@google.com>
Reviewed-by: Colin Nelson <colnnelson@google.com>