[rfc] Kernel Assisted CPU Profiling
This document proposes a design for CPU profiling in Fuchsia.

Concretely, it outlines a time-based, single-session CPU profiler that
utilizes frame pointers to obtain backtraces from user threads. Readers
familiar with profiling on Fuchsia may notice that the proposed design
is very similar to the experimental profiler that is currently checked
in, but it does contain some key differences.

Bug: 377907138
Bug: 423912552
Bug: 423914839
Change-Id: Ia66515a8d2476316ddf73bb0ae2b65dbf1b020f0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1277455
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>
Reviewed-by: Cameron Finucane <eep@google.com>
Reviewed-by: Dave Moore <davemoore@google.com>