[driver_runtime] Forced async cancel option on driver channel wait
This adds a mechanism to force channel wait cancellation to happen
asynchronously, even if the wait is started on a synchronized
dispatcher.

This will be used to support safe cancellation in the
rust bindings while the fuchsia-async dispatcher hack is still
in place, since the wait is held from the fuchsia-async dispatcher
but initiated on the driver's main dispatcher (usually).

Minor changes are made to the C++ api to separately control the
read and wait options, as previously that API passed the read
options to wait, which is kind of redundant since those options
are passed in through the read struct argument anyways. I assume
the intention of adding a separate options argument to the
wait call was intended to support wait-specific options.

Bug: 439021012
Change-Id: Ic5618341db0396a598b261a0c72e4f988284fb6d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1370803
API-Review: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>