[dhcp-client] Allow duplicate options
Update the parsing logic of DHCP messages to allow options to be
duplicated.

If the option has list semantics, we now merge the separate instances of
the options into a single cohesive list. If the option does not have
list semantics, it's generally nonsensical to attempt to merge the
separate instances. Instead, we use the first instance and ignore future
instances.

The original logic had a blanket statement that rejected the message if
any option is repeated. This leads to compatibility issues with existing
DHCP server implementations in the wild.

Multiply: dhcp-client-integration-test
Bug: 445178160
Change-Id: I4ba154a1e440c5d9e002fe4f195bd733cbb4adfe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1368976
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>