[netlink][tests] Simplify netlink tests
As noted in the parent CL, the backup routes are one of the reasons
tests were flaky. These routes are no longer being added in prod
environment where starnix by default uses the netstack marks and
suppresses the backup. This change simplifies the tests as well as
provide a more accurate coverage for what is supposed to be going on
in the non-test code paths.

Multiply: netlink-integration-tests
Bug: 438595958, 337297829, 410631890
Change-Id: I6a6a6964b127c16a71a60271c8a98d1cf74e039e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1371401
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Zeling Feng <zeling@google.com>