[starnix][test] Add a slot to inject a UTC clock handle for tests.
We want to verify that the linux syscalls do the right thing on
timerfd_* and read when used with TFD_TIMER_CANCEL_ON_SET, including the
scenarios where the UTC timeline changes. But, to do this, we also need
to call settimeofday from within a syscall test to elicit an UTC
timeline change, which is forbidden by default. This gadget would allow
that utc time adjustment.

Testing alarm management code which relies on changing the UTC timeline
requires the ability to inject a mutable clock into Starnix syscall
tests. This will only ever be used in tests. In production code,
external processes manage the UTC clock.

We reuse an existing FIDL API fuchsia.time/Management for this purpose,
although strictly speaking it is made for writable handles, and we do
not need handle writability, only to be able to send a
fuchsia.time/Adjust call.

Bug: 445541864
Change-Id: Idca758fe8b64e8eb1ca2f925df72fef29d9ef70e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1370201
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>