[build] Merge vmstat and ifconfig traces with main build trace
This merges the vmstat and ifconfig traces with that for the
main build, producing a trace that includes CPU, process, and
network data.

The result has some time-skew in it, which can be as much as
however long it takes Ninja to run GN, if a regeneration is
needed.

Change-Id: Ifa88707e5a3b51d6f568b817447d07594c7c7fa7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1370810
Reviewed-by: David Fang <fangism@google.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Commit-Queue: Aaron Wood <aaronwood@google.com>