[libc] Add support for supplying vmar to map library into
When loading a dynamic library via dlopen_vmo, the vmo ends up being
mapped into a child vmar of the process's root vmar. This makes it
challenging to apply scheduler profiles to alter the rate at which the
libraries' pages may be paged back to disk. In the case of a driver,
this means they are restricted to apply the scheduler profile to all
drivers in the process rather than just a specific driver. This change
provides a mechanism to target a specific driver by pre-carving out a
child vmar and loading the driver dso into it.

Change-Id: I25600cc5cb26d3383a8b81e7576c96f39e67028c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1360557
API-Review: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Suraj Malhotra <surajmalhotra@google.com>