[fxfs] Filter profiles to remove files that use external crypt keys.
Since the keys may not be present during replay they are removed for
now. They are filtered afterwards because during open time the keys are
fetch in the background as an optimization. Awaiting those keys to check
then would reduce those benefits.

Also moves the open recording into the `FxFile` impl, having it in
precise branches of the `FxDirectory` impl is a little unexpected and
likely harder to preserve.

Bug: 444217287
Change-Id: If04300ecd0cc9968197511baef0eef1208a56f35
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1368776
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>