[stack-migration] Improve "Rollback" reporting
Improve the reporting of the "Rollback" state to Cobalt.

If a device was automatically rolled back (e.g. its current boot is
`ForceNetstack2`), and it then observes the automated setting switch
from NS3 to NS2, stop reporting that the device is "RolledBack".

The original logic would leave the device perpetually reporting
"RolledBack", which obscures the cobalt analytics about the population.
It's more useful to classify such devices as "NotStarted".

Change-Id: I7365d244672a50b504ec84d60a19758995036791
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1371961
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Peter DeLong <pdelong@google.com>