[sestarnix] Apply boot state on audit logger
The kernel command line can fully disable or fully enabled the audit
framework using `audit=` option. If it is missing, the audit framework
is partially enabled and can print messages to the syslog, but will not
retain any messages for an audit-sink that is registered after the
messages are printed.

If the audit framework is fully enabled at boot, the messages are
printed to syslog, the backlog is enabled and can store at most
`backlog_limit` audit messages before dropping any other incoming
messages.

Bug: 438671380
Change-Id: I2d5307c4048d5b5928f951c91d0d5a6b1a57321b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1371843
Reviewed-by: Wez <wez@google.com>
Commit-Queue: Rareş Constantin <rarescon@google.com>