[netstack3] Consider ARP replies as override
The NUD state machine in Netstack3 has unified logic for both IPv4 and
IPv6; it attempts to implement the specification in RFC 4861 section 7
using ARP for IPv4. There are some impedance mismatches, however. One is
that ICMPv6 Neighbor Advertisements have an Override flag, which allows
a host to indicate whether it is intentionally updating its link-layer
address that may be cached by its neighbors. ARP does not have an
equivalent mechanism.

As a result, the current implementation specifies `false` for the value
of the Override flag in all cases for IPv4. This caused an unfortunate
result where, if a neighbor updates its link-layer address and replies
to future ARP requests with the new address, Netstack3 ignores the
confirmation because it was not considered to be an "override"
confirmation.

Instead, specify `true` for the Override flag for all ARP replies. This
will ensure that when the stack receives a reply with an updated
address, the neighbor entry's link-layer address will always be updated,
and if the reply was not unicast to us, the entry will be moved to the
STALE state. This should allow the stack to recover in the above
scenario, and happens to match Linux's behavior relatively closely.

Bug: 445196015
Fixed: 446189193
Test: netstack3-core-ip-integration-test
Change-Id: Ia52f064e8402e1de329a4f0ae8e1abf489017890
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1372132
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>