[assembled_system] Skip signing if in 'test ramdisk' mode
Per the mode's semantics, we should not be attempting to sign the
provided ramdisk when in test-ramdisk mode. We prevent this from
happening by enforcing mutual exclusivness in the assembled_system()
template and the tool, since it's straightforward enough there for the
user to not select both, and by skipping signing in the tool itself in
that mode, since there isn't an affirmative command-line directive
(outside of general board configuration) to similarly assert on.

Bug: 432535854
Change-Id: I31c222c6849f15330f93391f20bddaba77f33761
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1371945
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>