[fxfs] Replace volumes atomically when installing
Allow replacing an existing volume atomically when installing a new
volume. Previously, volume installation relied on any existing volumes
being deleted first. For eventual integration with fshost and Fuchsia
platform code, we provide a new `RootVolume::install_volume` function.

As part of this change, a logic bug was discovered in fsck that
disallowed both an EndFlush and DeleteVolume mutation in the same
transaction. This has since been fixed to allow both mutations as long
as they do not apply to the same object.

Bug: 441562849
Test: fx test fxfs-tests --test-filter="*install*"
Change-Id: I4c13cdcea0be22542acf47bdce0a73d5b685f5d1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1357164
Commit-Queue: Brandon Castellano <bcastell@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: Chris Suter <csuter@google.com>