[kernel][phys] PhysHandoffPtr is always const
It's already the case that all the instantiations used const
types.  Nothing but the PhysHandoff object itself is actually
modified (modulo a const_cast use for `k setopt`).  Formalize
this always-const property in the types.  In future it might make
sense for such memory to be read-only in the kernel.

Change-Id: I5ed993df8153325ceeb7c51692938a565bb55705
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1372282
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Roland McGrath <mcgrathr@google.com>