[starnix] Handcraft CFI expression to recover PC in the unwinder
This guarantees that a DW_CFA_extended_offset_sf directive is generated
for PC for the first restricted mode stack frame. This seems to have
been a longstanding issue with the the restricted mode integration with
the unwinder but wasn't noticed because it was working well enough.

Bug: 395606674
Bug: 416040479
Change-Id: I6a6a6964c8053a78789f3ec0e281038b1ce63385
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1364691
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>