[rcu] Introduce RcuList
RcuList is a doubly linked list that uses the Read-Copy-Update (RCU)
concurrency model. The list can be read and iterated without blocking.
Modifying the list requires external synchronization. I haven't figured
out the best way to ensure that the external synchronization is done
correctly. We'll need to solve that problem as we work with this code.

This CL is building towards using RCU for the PidTable in Starnix. In
order to support the PidTable, we need an RcuHashMap. In order to build
an RcuHashMap, we need an RcuList for the collision chains in the hash
map. This CL contains the RcuList.

Bug: b/445230228
Change-Id: Id9f1515bf39412b8d0a8c367d389023936ba16f9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1370768
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Owners-Override: Benjamin Lerman <qsr@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Adam Barth <abarth@google.com>