[unwinder] Refactor ARM EHABI data fetching
This is a preparatory change to support fetching longer data streams
from .ARM.extab sections, which is used for unwinding instructions that
are too long to fit in the Compact (Inline) model.

Now all unwinding opcodes are parsed together regardless of length and
collected into a byte vector that is constructed in execution order so
that the opcodes that span multiple bytes don't have to do any data
fetching, which will also help make the async cases easier to build in
the future.

Bug: 395606674
Bug: 416040479
Change-Id: I6a6a69646ccf0b0f6a6c5667275a17e77b8712b4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1371883
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>