[rcu] Introduce RcuPtrRef
This type tracks the lifetime of the RcuReadScope in which the pointer
was read. This type lets us remove RcuReadScope::as_ref, which was
dangerous because there was no way to ensure that the pointer being
converted to a reference was actually managed by RCU. Now, we mint
RcuPtrRef during RcuPtr::read, which means we can track their lifetime
safely.

Bug: b/445230228
Change-Id: Iba8210625443c99c41bf7a35088b5ac58b79efb4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1372231
Reviewed-by: Adam Perry <adamperry@google.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Commit-Queue: Adam Barth <abarth@google.com>