[build][idk] Disallow public* parameters if no_headers
Do not allow `public_configs`, or `public_deps` to be specified for
prebuilt libraries if `no_headers` is true. This replaces logic that
handled `public_deps` as `deps` if `no_headers` was true.
Also require that `public` is an empty list to prevent incusion of
headers specified in `sources`.

Change-Id: I5d343c4972d491ead5b8e2692feed1c8ed2a3f39
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1360645
Commit-Queue: David Dorwin <ddorwin@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>