[resultstore] ResultStore config and wrappers
Prepare configuration files for use of resultstore.
The -gcertauth variant is intended for users in a Corp
environment that is capable of exchanging LOAS for OAuth2
tokens using a credentials helper (from reclient).
The fallback is to use Application Default Credentials.

rsninja.sh is a drop-in replacement that adds resultstore-enabling
options to ninja when it detects that is wrapped by rsproxy-wrap.sh.
Fallback behavior is just the original ninja invocation.

These files are not integrated into the build yet.

Bug: 390427892
Change-Id: I50f2acdcacb80e30054636088bd6899329a9c9e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1369085
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Commit-Queue: David Fang <fangism@google.com>
Reviewed-by: David Turner <digit@google.com>